addTablComponent
