+++ src/assets/css/utils/_vars.scss
@@ -23,7 +23,7 @@
 }
 
 //font
-$font-size-body: 24px;
+$font-size-body: 16px;
 
 // transition
 $duration: 250ms;
