--- src/pages/DashboardPage/Dashboard.scss
@@ -1,3 +1,0 @@
-// .few {
-//   color: red;
-// }
